fix(ingestion/dbt): address contract ingestion review findings
Extract contract unit tests, probe the Discovery API in test_connection, fail incomplete pagination, disambiguate unnamed native-constraint URNs, and use the canonical per-dataset contract URN.
